Adds test for parsing unmatched quotes (#11)
