+++ test/fixture/nested_html_formatted.html.erb
@@ -1,6 +1,7 @@
 <div class="card">
   <span class="small">
     <%= t(".title") + " " + t(".description") + " " + t(".pretty_long") %>
-    '"This is a quote within a quote"'
+    'This is a single quote'
   </span>
+  "This is a double quote"
 </div>
