+++ test/fixture/nested_html_unformatted.html.erb
@@ -1,1 +1,1 @@
-<div class='card'><span class='small'><%= t(".title") + " " + t(".description") + " " + t(".pretty_long") %>'"This is a quote within a quote"'</span></div>
+<div class='card'><span class='small'><%= t(".title") + " " + t(".description") + " " + t(".pretty_long") %>'This is a single quote'</span>"This is a double quote"</div>
